4.15 - 4.22 Add users, login & token auth

+++ controllers/blogs.js
@@ -1,8 +1,12 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', {username: 1, name: 1})
     response.json(blogs)
 })
 
@@ -19,19 +23,23 @@ blogsRouter.get('/:id', async (request, response, next) => {
     }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', middleware.userExtractor, async (request, response, next) => {
     const body = request.body
+
+    const user = request.user
 
     const blog = new Blog({
         title: body.title,
         author: body.author,
+        user: user.id,
         url: body.url,
         likes: body.likes || 0
     })
 
     try {
         const savedBlog = await blog.save()
-        console.log(savedBlog)
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
         response.status(201).json(savedBlog)
     } catch (error) {
         next(error)
@@ -39,13 +47,22 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-    try {
-        await Blog.findByIdAndDelete(request.params.id)
-        response.status(204).end()
-      } catch(error) {
-        next(error)
-      }
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response, next) => {
+    const user = request.user
+
+    const blog = await Blog.findById(request.params.id)
+    console.log('=>', blog.user.toString())
+
+    if (blog.user.toString() === user.id.toString()) {
+        try {
+            await Blog.findByIdAndDelete(request.params.id)
+            response.status(204).end()
+          } catch(error) {
+            next(error)
+          }
+    } else {
+        return response.status(401).json({ error: 'unauthority to delete blog'})
+    }
     })
 
 blogsRouter.put('/:id', async (request, response, next) => {
